requirements_minimal.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
paho-mqtt==1.6.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
jinja2==3.1.2
python-json-logger==2.0.7
